Conflicts:
	app/models/attachment.rb
	app/models/comment.rb
	spec/models/discussion_spec.rb

+++ app/models/attachment.rb
@@ -4,6 +4,8 @@ class Attachment < ActiveRecord::Base
 
   validates_presence_of :filename, :location, :user_id
 
+  before_destroy :delete_from_storage
+
   def is_an_image?
     %w[jpg jpeg png gif].include?(filetype)
   end
@@ -20,12 +22,22 @@ class Attachment < ActiveRecord::Base
     filename.split('.').last.downcase
   end
 
-  def delete_attachment
-    storage = Fog::Storage.new({aws_access_key_id: ENV['AWS_ACCESS_KEY_ID'], aws_secret_access_key: ENV['AWS_SECRET_ACCESS_KEY'], provider: 'AWS'})
+  def delete_from_storage
+    storage = Fog::Storage.new({aws_access_key_id: ENV['AWS_ACCESS_KEY_ID'],
+                                aws_secret_access_key: ENV['AWS_SECRET_ACCESS_KEY'],
+                                provider: 'AWS'})
+
     bucket = storage.directories.get(ENV['AWS_ATTACHMENTS_BUCKET'])
-    file = bucket.files.get(URI.parse(URI.encode(self.location)).path)
-    file.destroy
+    filename = URI.decode(URI(URI.encode(self.location)).path).gsub(/^\//, '')
+    file = bucket.files.get(filename)
+
+    if file
+      file.destroy
+    else
+      raise "attachment filename not found within bucket: #{filename}"
+    end
   end
+
 end
